Prelim frame work
